Results_NB
This NB shows how to get the results of the analysis we did in the study. 
Note that this notebook uses resting-state data to showcase this. 
If you want to do it for task data, the code is identical, but please remember to change the paths to the corresponding task data.